requirements.txt
pandas
numpy
joblib
scikit-learn
imblearn
matplotlib
seaborn
rdkit-pypi
scipy
